[xarray-beam] Fix a bug in zarr._validate_chunk, which failed with KeyError when an offset is not specified for a dimension which has coordinates in both the template and the chunk.
It should instead assume offset 0 for dimensions where an offset isn't specified, and it does already do this for dimensions which don't have coordinates/an index.

This required updating offsets_to_slices to return a slice for all dimensions where the chunk size is known, defaulting to slice(0, size, 1).

test_2d_chunks_to_zarr has been updated to add a case with separate coordinates for the two dimensions, which is sufficient to trigger the issue and fails before the fix.

Renames zarr._validate_chunk to _validate_zarr_chunk to avoid confusion with core.validate_chunk.

PiperOrigin-RevId: 395055847