Use RangeSource inside a new ReadDatasets ptransform
This is equivalent to the existing xbeam.DatasetToChunks, except it uses the beam source API instead of writing custom PTransforms. Hopefully this will be a bit more efficient!

I have not yet exposed this as a public API because I want to test it more, and have the freedom to adjust the API. Probably I'll hook it up to xbeam.Dataset.from_xarray() initially.

PiperOrigin-RevId: 828074171